test: add failing test and change README
This is the starting point for extending C# with Package Templates. So
far i have added the folder "PackageTemplate" to
CSharpCompilerSyntaxTestProject and added the class
"PackageTemplateSyntaxTest.cs" to this. Running the unit tests for the
parser/lexer also runs this class, this is a good starting point for
TDD.

Also changed README.md to describe this repo, link to the original
is found in the README.

TODO: Create tests that match the PackageTemplate Syntax, start simple
and after making the tests go green, carry on to more complex versions