test: add failing test
Add folder "PackageTemplate" to CSharpCompilerSyntaxTestProject, add
class "PackageTemplateSyntaxTest.cs" to this folder and ensure that it
runs along with the other tests, successful! This is a good starting
point.

TODO: Create tests that match the PackageTemplate Syntax, start simple
and after making the tests go green, carry on to more complex versions