feature: parser now parses classes within templates
This needs to be worked with. At the moment, the original class parser
is used to parse the class within the template, this allows for a lot of
classes that aren't allowed within a template. To make this correct a
custom class parser must be implemented that can parse a class according
to template rules, but to be able to continue i made a short-cut and
plan to revisit when inst-statement parsing is done.