feat: add TemplateDeclaration, ITemplateSymbol, and TemplateSymbol
TemplateDeclaration is derived from SingleNamespaceOrTypeDeclaration and
has the most basic possible implementation to start with, after
discovering what features are needed within this class, they will be
implemented.

ITemplateSymbol is an interface derived from INamespaceOrTypeSymbol,
this interface adds nothing, it's just there to comply with the types
for a few methods

TemplateSymbol is derived from NamespaceOrTypeSymbol and has an
incomplete implementation, almost all methods are marked with
NotYetImplemented, and I hope that further testing will uncover what
methods are used and in what situation.