feature: add MergedTemplateDeclaration
Both Types and Namespaces have a "Merged" version of their declaration,
the reason for this is that they can be split over multiple files, eg.
partial classes, or namespaces that are declared in multiple files. This
is not the case for Templates (at least not now), so the
MergedTemplateDeclaration class is simply just a wrapper around a
TemplateDeclaration, the reason for this is that we need a subclass of
MergedNamespaceOrTypeDeclaration to conform to the parameters in certain
methods in the compiler.